requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
geoalchemy2==0.14.3
psycopg2-binary==2.9.9
pydantic==2.5.3
feedparser==6.0.10
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
lxml==5.1.0
httpx==0.27.0

